refactor: move coroutine configuration in a separate file
