+++ core/src/test/kotlin/com/rudderstack/kotlin/sdk/Utils.kt
@@ -9,6 +9,7 @@ import com.rudderstack.kotlin.sdk.internals.policies.DEFAULT_FLUSH_INTERVAL_IN_M
 import com.rudderstack.kotlin.sdk.internals.utils.empty
 import io.mockk.every
 import io.mockk.mockk
+import io.mockk.mockkStatic
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.TestDispatcher
 import kotlinx.coroutines.test.TestScope
@@ -17,12 +18,15 @@ import java.io.BufferedReader
 const val ANONYMOUS_ID = "<anonymous-id>"
 
 fun mockAnalytics(testScope: TestScope, testDispatcher: TestDispatcher): Analytics {
-    val mock = mockk<Analytics>(relaxed = true)
-    every { mock.analyticsScope } returns testScope
-    every { mock.analyticsDispatcher } returns testDispatcher
-    every { mock.storageDispatcher } returns testDispatcher
-    every { mock.networkDispatcher } returns testDispatcher
-    return mock
+    val mockAnalytics = mockk<Analytics>(relaxed = true)
+
+    mockkStatic("com.rudderstack.kotlin.sdk.CoroutineExtensionsKt")
+    every { analyticsScope } returns testScope
+    every { analyticsDispatcher } returns testDispatcher
+    every { storageDispatcher } returns testDispatcher
+    every { networkDispatcher } returns testDispatcher
+
+    return mockAnalytics
 }
 
 fun Message.applyMockedValues() {
